+++ src/main/java/act/cli/meta/CommanderClassMetaInfoManager.java
@@ -24,16 +24,12 @@ import static act.Destroyable.Util.destroyAll;
 
 import act.asm.Type;
 import act.util.LogSupportedDestroyableBase;
-import org.osgl.logging.LogManager;
-import org.osgl.logging.Logger;
 
 import java.util.*;
 import javax.enterprise.context.ApplicationScoped;
 
 @ApplicationScoped
 public class CommanderClassMetaInfoManager extends LogSupportedDestroyableBase {
-
-    private static final Logger logger = LogManager.get(CommanderClassMetaInfoManager.class);
 
     private Map<String, CommanderClassMetaInfo> commands = new HashMap<>();
     private Map<Type, List<CommanderClassMetaInfo>> subTypeInfo = new HashMap<>();